Changes food rating to type float in model class FoodEntry to work with RatingBar component.

+++ FoodieApp/app/src/main/java/com/versilistyson/foodieapp/FoodEntry.kt
@@ -4,14 +4,14 @@ import android.net.Uri
 import java.io.Serializable
 
 class FoodEntry(id:Int,restaurant_name:String,restaurant_type:String,item_name:
-String,photoUri:Uri?,food_rating:Int,comments:String,price:String,date_of_visit:String):Serializable{
+String,photoUri:Uri?,food_rating:Float,comments:String,price:String,date_of_visit:String):Serializable{
 
     val id=id
     var restaurant_name=restaurant_name
     var restaurant_type=restaurant_type
     var item_name=item_name
     val photoUri=photoUri
-    var food_rating=food_rating
+    var food_rating = food_rating
     var comments=comments
     var price=price
     var date_of_visit=date_of_visit
